+++ sickbeard/helpers.py
@@ -1152,8 +1152,13 @@ def getURL(url, post_data=None, params=None, headers=None, timeout=30, session=N
                 }
 
         # decide if we get or post data to server
+        if 'post_json' in kwargs:
+            kwargs.setdefault('json', kwargs.get('post_json'))
+            del(kwargs['post_json'])
         if post_data:
-            resp = session.post(url, data=post_data, timeout=timeout, **kwargs)
+            kwargs.setdefault('data', post_data)
+        if 'data' in kwargs or 'json' in kwargs:
+            resp = session.post(url, timeout=timeout, **kwargs)
         else:
             resp = session.get(url, timeout=timeout, **kwargs)
 
@@ -1169,26 +1174,35 @@ def getURL(url, post_data=None, params=None, headers=None, timeout=30, session=N
             return
 
     except requests.exceptions.HTTPError as e:
-        logger.log(u'HTTP error %s while loading URL %s' % (e.errno, url), logger.WARNING)
+        logger.log(u'HTTP error %s while loading URL %s' % (e.errno, e.request.url), logger.WARNING)
         return
     except requests.exceptions.ConnectionError as e:
-        logger.log(u'Internet connection error msg:%s while loading URL %s' % (str(e.message), url), logger.WARNING)
+        if not kwargs.get('mute_connect_err'):
+            logger.log(u'Connection error msg:%s while loading URL %s' % (e.message, e.request.url), logger.WARNING)
         return
     except requests.exceptions.ReadTimeout as e:
-        logger.log(u'Read timed out msg:%s while loading URL %s' % (str(e.message), url), logger.WARNING)
+        logger.log(u'Read timed out msg:%s while loading URL %s' % (e.message, e.request.url), logger.WARNING)
         return
     except (requests.exceptions.Timeout, socket.timeout) as e:
-        logger.log(u'Connection timed out msg:%s while loading URL %s' % (str(e.message), url), logger.WARNING)
+        logger.log(u'Connection timed out msg:%s while loading URL %s'
+                   % (e.message, hasattr(e, 'request') and e.request.url or url), logger.WARNING)
         return
     except Exception as e:
+        url = hasattr(e, 'request') and e.request.url or url
         if e.message:
-            logger.log(u'Exception caught while loading URL %s\r\nDetail... %s\r\n%s' % (url, str(e.message), traceback.format_exc()), logger.WARNING)
+            logger.log(u'Exception caught while loading URL %s\r\nDetail... %s\r\n%s'
+                       % (url, e.message, traceback.format_exc()), logger.WARNING)
         else:
-            logger.log(u'Unknown exception while loading URL %s\r\nDetail... %s' % (url, traceback.format_exc()), logger.WARNING)
+            logger.log(u'Unknown exception while loading URL %s\r\nDetail... %s'
+                       % (url, traceback.format_exc()), logger.WARNING)
         return
 
     if json:
-        return resp.json()
+        try:
+            return resp.json()
+        except (TypeError, Exception) as e:
+            logger.log(u'JSON data issue from URL %s\r\nDetail... %s' % (url, e.message), logger.WARNING)
+            return None
 
     return resp.content